BestSeller 목록

+++ src/main/java/com/forest/books/domain/ItemView.java
@@ -23,7 +23,7 @@ public class ItemView {
 	private String description; // 상품설명(요약)
 	private int priceSales; // 판매가
 	private int priceStandard; // 정가
-	private int customerReviewRank; // 회원 리뷰 평점(별점 평균) : 0~10점(별0.5개당 1점)
+	private double customerReviewRank; // 회원 리뷰 평점(별점 평균) : 0~10점(별0.5개당 1점)
 	private Integer bestRank; // (베스트셀러인 경우만 노출) 베스트셀러 순위 정보 - 정수
 	
 	// TODO: 중고도서 정보 - 있는지 없는지, 중고가, 해당 중고도서 상세 링크
